+++ src/actors/shipUtils.ts
@@ -1,4 +1,4 @@
-import { Actor, vec } from "excalibur";
+import { Actor, Engine, vec } from "excalibur";
 
 export const flyInRandomDirection = (actor: Actor) => {
   actor.vel = vec(
@@ -25,4 +25,30 @@ export const itsBeenAFewSeconds = (timeStarted: Date) => {
   const timeDiff = now - timeStarted.getTime();
 
   return timeDiff > 1 * ONE_SECOND;
+};
+
+export const bounceOffEdges = (actor: Actor, game: Engine) => {
+  // If the ball collides with the left side
+  // of the screen reverse the x velocity
+  if (actor.pos.x - 20 < actor.width / 2) {
+    actor.vel.x *= -1;
+  }
+
+  // If the ball collides with the right side
+  // of the screen reverse the x velocity
+  if (actor.pos.x > game.drawWidth + 20) {
+    actor.vel.x *= -1;
+  }
+
+  // If the ball collides with the top
+  // of the screen reverse the y velocity
+  if (actor.pos.y < -20) {
+    actor.vel.y *= -1;
+  }
+
+  // If the ball collides with the bottom
+  // of the screen reverse the y velocity
+  if (actor.pos.y - 20 > game.drawHeight) {
+    actor.vel.y *= -1;
+  }
 };
