+++ src/utils/createMachine.ts
@@ -17,7 +17,7 @@ export type State<S> = {
 type Transition<S> = {
   [key: string]: {
     nextState: State<S>;
-    effect?: () => void;
+    effect: () => void;
   };
 };
 
@@ -41,40 +41,12 @@ export function createMachine<S>(
       transitionName: ShipTransitions
     ) {
       const { states } = stateMachineDefinition;
-
-      /**
-       * Was a valid state passed in?
-       */
       const currentStateDefinition = states[currentState.type];
-      if (!currentStateDefinition) {
-        console.warn(`Unknown state ✨${currentState}✨`);
-        return currentState;
-      }
 
-      /**
-       * Does this state allow this transition
-       */
       const transition = currentStateDefinition.transitions[transitionName];
-      if (!transition) {
-        console.warn(
-          `You can not 🚙 ${transitionName} 🚙 from the ✨ ${currentState.type} ✨ state`
-        );
-        return currentState;
-      }
-
-      /**
-       * Effects on transitions run before we move to the next state
-       */
-      transition.effect && transition.effect();
-
-      /**
-       * Log a time stamp of the transition so the next state has a start time
-       */
+      transition.effect();
       transition.nextState.at = new Date();
 
-      /**
-       * Transition to the next state
-       */
       machine.value = transition.nextState;
       return transition.nextState;
     }
